Pull parse state classes out of GlslParser.
