+++ src/commonMain/kotlin/baaahs/gl/glsl/parser/Macro.kt
@@ -1,0 +1,6 @@
+package baaahs.gl.glsl.parser
+
+class Macro(
+    val params: List<Token>?,
+    val replacement: List<Token>
+)
