+++ src/commonMain/kotlin/baaahs/gl/glsl/parser/State.kt
@@ -1,4 +1,4 @@
-package baaahs.gl.glsl
+package baaahs.gl.glsl.parser
 
 interface State<S: State<S>> {
     fun visit(token: Token): S
